move settings, celery, wsgi, etc. to conf app
