+++ ohmg/conf/wsgi.py
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ohmg.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ohmg.conf.settings")
 
 application = get_wsgi_application()
